Add files via upload
Task 2 - User Engagement Analysis
As telecom brands are the data providers of all online activities, meeting user requirements, and creating an engaging user experience is a prerequisite for them. Building & improving the QoS (Quality of Service) to leverage the mobile platforms and to get more users for the business is good but the success of the business would be determined by the user engagement and activity of the customers on available apps. 

In telecommunication, tracking the user activities on the database sessions is a good starting point to appreciate the user engagement for the overall applications and per application as well. If we can determine the level of engagement of a random user for any application, then it could help the technical teams of the business to know where to concentrate network resources for different clusters of customers based on the engagement scores.

In the current dataset you’re expected to track the user’s engagement using the following engagement metrics: 
sessions frequency 
the duration of the session 
the session total traffic (download and upload (bytes))

Task 2.1 - Based on the above submit the Python script and slide:
Aggregate the above metrics per customer id (MSISDN) and report the top 10 customers per engagement metric 
Normalize each engagement metric and run a k-means (k=3) to classify customers into three groups of engagement. 
Compute the minimum, maximum, average & total non-normalized metrics for each cluster. Interpret your results visually with accompanying text explaining your findings.
Aggregate user total traffic per application and derive the top 10 most engaged users per application
Plot the top 3 most used applications using appropriate charts.  
Using the k-means clustering algorithm, group users in k engagement clusters based on the engagement metrics: 
What is the optimized value of k (use the elbow method for this)?  
Interpret your findings.